frontend/requirements.txt
python-vlc
pyside6
